contact page split into components
